app/requirements.txt
Flask
Flask-MySQL
MySQL
mysql-connector
Flask-WTF
Flask-Passlib
